+++ src/utils/formatDuration.ts
@@ -3,6 +3,9 @@ const om = 60 * os
 const oh = 60 * om
 
 export const formatDuration = (ms: number) => {
+  if (isNaN(ms) || ms <= 0) {
+    return '00:00'
+  }
   const h = Math.floor(ms / oh)
   ms %= oh
   let m: number | string = Math.floor(ms / om)